# docs/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: HSE Platform API
  description: "\nHSE platform API — Phase 0 Foundation (spec `docs/specs/0-foundation.md` v1.0).\n\n\
    * Auth: `POST /api/v1/auth/login` sets the httpOnly SameSite=Lax cookie `hse_session` (JWT) and\n\
    \  returns `{access_token, user}`. Send the cookie or `Authorization: Bearer <token>`.\n* Errors:\
    \ `{\"detail\": {\"code\", \"message\", \"message_ar\", \"errors\"}}`; 401 unauthenticated,\n  403\
    \ forbidden, 404 not found / out of scope, 409 conflict / invalid transition, 422 validation.\n* Lists:\
    \ `?page=1&page_size=50` → `{items, total, page, page_size}`.\n* Timestamps are UTC; display and day\
    \ boundaries use the project timezone (Asia/Riyadh).\n"
  version: 0.1.0
paths:
...
  /api/v1/projects:
    get:
      tags:
      - projects
      summary: List projects visible to the caller
      operationId: list_projects
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              $ref: '#/components/schemas/ProjectStatus'
          - type: 'null'
          title: Status
      - name: project_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ProjectType'
          - type: 'null'
          title: Project Type
      - name: q
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          description: Code/name search.
          title: Q
        description: Code/name search.
      - name: sort
        in: query
        required: false
        schema:
          enum:
          - code
          - -code
          - name
          - -name
          - start_date
          - -start_date
          type: string
          default: code
          title: Sort
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: 1-based page number.
          default: 1
          title: Page
        description: 1-based page number.
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Items per page.
          default: 50
          title: Page Size
        description: Items per page.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPage'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
      - projects
      summary: Create a project (HSE Manager); starts in planning with default settings
      operationId: create_project
      security:
      - sessionCookie: []
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}:
    get:
      tags:
      - projects
      summary: Get a project
      operationId: get_project
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - projects
      summary: Update a project (HSE Manager)
      operationId: update_project
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}/transitions:
    post:
      tags:
      - projects
      summary: Change project status (HSE Manager)
      operationId: transition_project
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectTransitionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/projects/{project_id}/settings:
    get:
      tags:
      - projects
      summary: Get project settings
      operationId: get_project_settings
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSettingsRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      tags:
      - projects
      summary: Update project settings (HSE Manager); audited as settings_changed
      operationId: update_project_settings
      security:
      - sessionCookie: []
      - bearerAuth: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectSettingsUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSettingsRead'
        '401':
          description: Not authenticated, session expired, or login rejected.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Missing capability, read-only role, contractor suspended, or privacy notice not
            acknowledged (PRIVACY_ACK_REQUIRED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Not found or outside the caller's project/contractor scope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Invalid state transition, closed project, duplicate value or business-rule conflict.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error (field errors in detail.errors).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DateFormatEn:
      type: string
      enum:
      - DD MMM YYYY
      - DD/MM/YYYY
      title: DateFormatEn
    DigitStyle:
      type: string
      enum:
      - western
      - arabic_indic
      title: DigitStyle
...
    ErrorCode:
      type: string
      enum:
      - UNAUTHENTICATED
      - INVALID_CREDENTIALS
      - ACCOUNT_LOCKED
      - SESSION_EXPIRED
      - PRIVACY_ACK_REQUIRED
      - PRIVACY_NOTICE_VERSION_MISMATCH
      - INVITE_INVALID
      - INVITE_EXPIRED
      - RESET_TOKEN_INVALID
      - WEAK_PASSWORD
      - CURRENT_PASSWORD_INCORRECT
      - FORBIDDEN
      - READ_ONLY_ROLE
      - CONTRACTOR_SUSPENDED
      - SELF_MODIFICATION_FORBIDDEN
      - ROLE_NOT_ASSIGNABLE
      - NOT_FOUND
      - VALIDATION_ERROR
      - DUPLICATE_VALUE
      - INVALID_TRANSITION
      - TRANSITION_CONDITION_NOT_MET
      - PROJECT_CLOSED
      - SITE_INACTIVE
      - ZONE_ARCHIVED
      - CONTRACTOR_NOT_APPROVED
      - LAST_HSE_MANAGER
      - SOD_CONFLICT
      - NOT_IMPLEMENTED
      - INTERNAL_ERROR
      title: ErrorCode
      description: 'Machine-readable error codes. Stable: the frontend translates these.'
    ErrorDetail:
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Stable machine code; map it to EN/AR UI text.
        message:
          type: string
          title: Message
          description: English, human-readable message.
        message_ar:
          anyOf:
          - type: string
          - type: 'null'
          title: Message Ar
          description: Arabic message, when available.
        errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldError'
            type: array
          - type: 'null'
          title: Errors
          description: Field-level errors (422 validation, 409 duplicates).
      type: object
      required:
      - code
      - message
      title: ErrorDetail
    ErrorResponse:
      properties:
        detail:
          $ref: '#/components/schemas/ErrorDetail'
      type: object
      required:
      - detail
      title: ErrorResponse
      description: Envelope for every 4xx/5xx response.
...
    FieldError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Loc
          description: Path to the field, e.g. ["body", "zone_type"].
        msg:
          type: string
          title: Msg
          description: English description of the problem.
        type:
          type: string
          title: Type
          description: Machine-readable error type, e.g. 'missing', 'value_error'.
      type: object
      required:
      - loc
      - msg
      - type
      title: FieldError
      description: One field-level problem (used with VALIDATION_ERROR and DUPLICATE_VALUE).
...
    HijriCalendar:
      type: string
      enum:
      - umm_al_qura
      title: HijriCalendar
...
    KpiBaseHours:
      type: integer
      enum:
      - 200000
      - 1000000
      title: KpiBaseHours
      description: Allowed KPI normalisation bases (§3.9, §5.5 rule 31).
    Language:
      type: string
      enum:
      - en
      - ar
      title: Language
...
    ProjectCreate:
      properties:
        code:
          type: string
          pattern: ^[A-Z0-9-]{3,12}$
          title: Code
          description: Unique, e.g. ANIA-EXP. Immutable.
        name_en:
          type: string
          maxLength: 150
          minLength: 1
          title: Name En
        name_ar:
          type: string
          maxLength: 150
          minLength: 1
          title: Name Ar
          description: Must contain Arabic script.
        project_type:
          $ref: '#/components/schemas/ProjectType'
        client_name_en:
          type: string
          maxLength: 150
          minLength: 1
          title: Client Name En
        client_name_ar:
          type: string
          maxLength: 150
          minLength: 1
          title: Client Name Ar
        city:
          type: string
          maxLength: 80
          minLength: 1
          title: City
        start_date:
          type: string
          format: date
          title: Start Date
        planned_end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Planned End Date
        airport_icao:
          anyOf:
          - type: string
            pattern: ^[A-Z]{4}$
          - type: 'null'
          title: Airport Icao
          description: Required iff project_type = airport; Saudi codes start with OE.
      additionalProperties: false
      type: object
      required:
      - code
      - name_en
      - name_ar
      - project_type
      - client_name_en
      - client_name_ar
      - city
      - start_date
      title: ProjectCreate
    ProjectPage:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ProjectRead'
          type: array
          title: Items
        total:
          type: integer
          title: Total
          description: Total number of matching items across all pages.
        page:
          type: integer
          title: Page
          description: 1-based page number.
        page_size:
          type: integer
          title: Page Size
          description: Items per page (max 200).
      type: object
      required:
      - items
      - total
      - page
      - page_size
      title: ProjectPage
    ProjectRead:
      properties:
        created_at:
          type: string
          format: date-time
          title: Created At
          description: UTC timestamp.
        updated_at:
          type: string
          format: date-time
          title: Updated At
          description: UTC timestamp.
        id:
          type: string
          format: uuid
          title: Id
        code:
          type: string
          title: Code
        name_en:
          type: string
          title: Name En
        name_ar:
          type: string
          title: Name Ar
        project_type:
          $ref: '#/components/schemas/ProjectType'
        is_airport:
          type: boolean
          title: Is Airport
          description: 'Derived: project_type = airport.'
        client_name_en:
          type: string
          title: Client Name En
        client_name_ar:
          type: string
          title: Client Name Ar
        city:
          type: string
          title: City
        start_date:
          type: string
          format: date
          title: Start Date
        planned_end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Planned End Date
        status:
          $ref: '#/components/schemas/ProjectStatus'
        status_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Reason
          description: Reason of the last status change, if any.
        airport_icao:
          anyOf:
          - type: string
          - type: 'null'
          title: Airport Icao
      type: object
      required:
      - created_at
      - updated_at
      - id
      - code
      - name_en
      - name_ar
      - project_type
      - is_airport
      - client_name_en
      - client_name_ar
      - city
      - start_date
      - planned_end_date
      - status
      - status_reason
      - airport_icao
      title: ProjectRead
    ProjectSettingsRead:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
        ltifr_base_hours:
          $ref: '#/components/schemas/KpiBaseHours'
          description: LTIFR normalisation base.
        rate_base_hours:
          $ref: '#/components/schemas/KpiBaseHours'
          description: Base for TRIR, DART, LTISR, other rates.
        ltifr_base_label_en:
          type: string
          title: Ltifr Base Label En
          examples:
          - per 1,000,000 h
        ltifr_base_label_ar:
          type: string
          title: Ltifr Base Label Ar
          examples:
          - لكل 1,000,000 ساعة
        rate_base_label_en:
          type: string
          title: Rate Base Label En
          examples:
          - per 200,000 h
        rate_base_label_ar:
          type: string
          title: Rate Base Label Ar
        timezone:
          $ref: '#/components/schemas/ProjectTimezone'
        show_hijri:
          type: boolean
          title: Show Hijri
        hijri_calendar:
          $ref: '#/components/schemas/HijriCalendar'
        default_language:
          $ref: '#/components/schemas/Language'
        week_start:
          $ref: '#/components/schemas/WeekStart'
        digits:
          $ref: '#/components/schemas/DigitStyle'
        date_format_en:
          $ref: '#/components/schemas/DateFormatEn'
        audit_retention_years:
          type: integer
          maximum: 10.0
          minimum: 1.0
          title: Audit Retention Years
        inactive_account_days:
          type: integer
          maximum: 365.0
          minimum: 30.0
          title: Inactive Account Days
        saved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Saved At
          description: First explicit save by an HSE Manager; required before planning→active.
        updated_at:
          type: string
          format: date-time
          title: Updated At
        updated_by_user_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Updated By User Id
      type: object
      required:
      - project_id
      - ltifr_base_hours
      - rate_base_hours
      - ltifr_base_label_en
      - ltifr_base_label_ar
      - rate_base_label_en
      - rate_base_label_ar
      - timezone
      - show_hijri
      - hijri_calendar
      - default_language
      - week_start
      - digits
      - date_format_en
      - audit_retention_years
      - inactive_account_days
      - saved_at
      - updated_at
      - updated_by_user_id
      title: ProjectSettingsRead
    ProjectSettingsUpdate:
      properties:
        ltifr_base_hours:
          anyOf:
          - $ref: '#/components/schemas/KpiBaseHours'
          - type: 'null'
        rate_base_hours:
          anyOf:
          - $ref: '#/components/schemas/KpiBaseHours'
          - type: 'null'
        timezone:
          anyOf:
          - $ref: '#/components/schemas/ProjectTimezone'
          - type: 'null'
        show_hijri:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Show Hijri
        hijri_calendar:
          anyOf:
          - $ref: '#/components/schemas/HijriCalendar'
          - type: 'null'
        default_language:
          anyOf:
          - $ref: '#/components/schemas/Language'
          - type: 'null'
        week_start:
          anyOf:
          - $ref: '#/components/schemas/WeekStart'
          - type: 'null'
        digits:
          anyOf:
          - $ref: '#/components/schemas/DigitStyle'
          - type: 'null'
        date_format_en:
          anyOf:
          - $ref: '#/components/schemas/DateFormatEn'
          - type: 'null'
        audit_retention_years:
          anyOf:
          - type: integer
            maximum: 10.0
            minimum: 1.0
          - type: 'null'
          title: Audit Retention Years
        inactive_account_days:
          anyOf:
          - type: integer
            maximum: 365.0
            minimum: 30.0
          - type: 'null'
          title: Inactive Account Days
      additionalProperties: false
      type: object
      title: ProjectSettingsUpdate
      description: Partial update (HSE Manager only, rule 30). An empty body still marks settings saved.
    ProjectStatus:
      type: string
      enum:
      - planning
      - active
      - on_hold
      - closed
      title: ProjectStatus
      description: Project lifecycle (§4.1).
    ProjectTimezone:
      type: string
      enum:
      - Asia/Riyadh
      title: ProjectTimezone
    ProjectTransitionRequest:
      properties:
        to_status:
          $ref: '#/components/schemas/ProjectStatus'
        reason:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Reason
      additionalProperties: false
      type: object
      required:
      - to_status
      title: ProjectTransitionRequest
      description: 'Allowed (§4.1): planning→active (≥1 site and settings saved), active→on_hold (reason),

        on_hold→active, active/on_hold→closed (reason), closed→active (reason, reopen).'
    ProjectType:
      type: string
      enum:
      - airport
      - building_highrise
      - infrastructure
      - industrial
      - other
      title: ProjectType
    ProjectUpdate:
      properties:
        name_en:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Name En
        name_ar:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Name Ar
        project_type:
          anyOf:
          - $ref: '#/components/schemas/ProjectType'
          - type: 'null'
        client_name_en:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Client Name En
        client_name_ar:
          anyOf:
          - type: string
            maxLength: 150
            minLength: 1
          - type: 'null'
          title: Client Name Ar
        city:
          anyOf:
          - type: string
            maxLength: 80
            minLength: 1
          - type: 'null'
          title: City
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        planned_end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Planned End Date
        airport_icao:
          anyOf:
          - type: string
            pattern: ^[A-Z]{4}$
          - type: 'null'
          title: Airport Icao
      additionalProperties: false
      type: object
      title: ProjectUpdate
      description: Partial update. Rejected with 409 PROJECT_CLOSED when the project is closed.
...
    WeekStart:
      type: string
      enum:
      - sunday
      - monday
      title: WeekStart
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      description: httpOnly SameSite=Lax cookie set by POST /auth/login.
      in: cookie
      name: hse_session
    bearerAuth:
      type: http
      description: 'Same JWT as the cookie, as `Authorization: Bearer <token>`.'
      scheme: bearer